Refactoring (for details see long message)
Refactoring quite a lot here on to make it easier to move forward with
keeping records of what has been delivered previously by:
 - moving a lot of the previous staging functionality into the
   delivery service
 - creating a runfolder service, which deals with many actions relating
   to runfolders. There is probably more code around which could benefit
   from being moved into here
