+++ delivery/handlers/project_handlers.py
@@ -1,7 +1,6 @@
 
 from delivery.handlers import *
 from delivery.handlers.utility_handlers import ArteriaDeliveryBaseHandler
-from delivery.repositories.project_repository import RunfolderProjectRepository
 
 
 class ProjectBaseHandler(ArteriaDeliveryBaseHandler):
@@ -11,8 +10,6 @@ class ProjectBaseHandler(ArteriaDeliveryBaseHandler):
 
     def initialize(self, **kwargs):
         self.runfolder_repo = kwargs["runfolder_repo"]
-        self.project_repo = RunfolderProjectRepository(
-            runfolder_repository=self.runfolder_repo)
         super(ProjectBaseHandler, self).initialize(kwargs)
 
 
@@ -34,7 +31,7 @@ class ProjectHandler(ProjectBaseHandler):
             ]
         }
         """
-        projects = list(self.project_repo.get_projects())
+        projects = list(self.runfolder_repo.get_projects())
         self.write_list_of_models_as_json(projects, key="projects")
