+++ delivery/repositories/project_repository.py
@@ -5,33 +5,6 @@ from delivery.services.file_system_service import FileSystemService
 from delivery.models.project import GeneralProject
 from delivery.exceptions import TooManyProjectsFound, ProjectNotFoundException
 
-class RunfolderProjectRepository(object):
-    """
-    Repository for materializing project instances
-    """
-
-    def __init__(self, runfolder_repository):
-        """
-        Instantiate a new repository
-        :param runfolder_repository: a `FileSystemBasedRunfolderRepository` or something the implements the
-        `get_runfolders` method
-        """
-        self.runfolder_repository = runfolder_repository
-
-    def get_projects(self):
-        """
-        Pick up all projects
-        :return: a generator of project instances
-        """
-        for runfolder in self.runfolder_repository.get_runfolders():
-            if runfolder.projects:
-                for project in runfolder.projects:
-                    yield project
-
-    def get_project(self, project_name):
-        for project in self.get_projects():
-            if project.name == project_name:
-                yield project
 
 class GeneralProjectRepository(object):
     """
