+++ delivery/services/staging_service.py
@@ -42,7 +42,6 @@ class StagingService(object):
                  runfolder_repo,
                  project_dir_repo,
                  project_links_directory,
-                 runfolder_project_repo,
                  session_factory):
         """
         Instantiate a new StagingService
@@ -53,7 +52,6 @@ class StagingService(object):
         :param project_dir_repo: a instance of GeneralProjectRepository
         :param project_links_directory: a path to a directory where links will be created temporarily
                                         before they are rsynced into staging (for batched deliveries etc)
-        :param runfolder_project_repo: A runfolder project repo
         :param session_factory: a factory method which can produce new sqlalchemy Session instances
         """
         self.staging_dir = staging_dir
@@ -62,7 +60,6 @@ class StagingService(object):
         self.runfolder_repo = runfolder_repo
         self.project_dir_repo = project_dir_repo
         self.project_links_directory = project_links_directory
-        self.runfolder_project_repo = runfolder_project_repo
         self.session_factory = session_factory
 
     @staticmethod
@@ -154,62 +151,6 @@ class StagingService(object):
             stage_order.status = StagingStatus.staging_failed
             session.commit()
             raise e
-
-    def _create_links_area_for_project_runfolders(self, project_name, projects):
-        """
-        Creates a directory in which it creates links to all runfolders for the projects
-        given. This is useful so that we can then rsync that directory to
-        the staging area.
-        :param project_name: name of the project
-        :param projects: runfolders with the specified project on them
-        :return: the path to the dir created
-        """
-
-        project_dir = os.path.join(self.project_links_directory, project_name)
-        try:
-            FileSystemService.mkdir(project_dir)
-        except FileExistsError as e:
-            log.warning("Project dir: {} already exists".format(project_dir))
-
-        for project in projects:
-            try:
-                link_name = os.path.join(project_dir, project.runfolder_name)
-                FileSystemService.symlink(project.path, link_name)
-            except FileExistsError:
-                log.warning("Project link: {} already exists".format(project_dir))
-                continue
-
-        return FileSystemService.abspath(project_dir)
-
-    def stage_runfolders_for_project(self, project_id, delivery_type):
-        projects = list(self.runfolder_project_repo.get_project(project_name=project_id))
-
-        if len(projects) < 1:
-            raise ProjectNotFoundException("Could not find any Project folders"
-                                           " for project name: {}".format(project_id))
-
-        # TODO
-        # If delivery type == "clean"
-        #   No runfolder project can have been delivered before
-        # If delivery type == "batch"
-        #   Only delivery none delivered runfolders
-        # If delivery type == force
-        #   Re-deliver all independent of status
-
-        # Link all <runfolder name>/Projects/<proj name>/<runfolder name> into same directory
-        # TODO Continue here...
-        links_directory = self._create_links_area_for_project_runfolders(project_name=project_id,
-                                                                         projects=projects)
-
-        # Stage
-        staging_order = self.staging_repo.create_staging_order(source=links_directory,
-                                                               status=StagingStatus.pending,
-                                                               staging_target_dir=self.staging_dir)
-        self.stage_order(staging_order)
-
-        # TODO Maybe, once stage is complete remove the dir..
-        return {project_id: staging_order.id}
-
 
     def create_new_stage_order(self, path):
         staging_order = self.staging_repo.create_staging_order(source=path,
