+++ tests/unit_tests/handlers/test_delivery_project_handlers.py
@@ -16,10 +16,21 @@ class TestProjectHandlers(AsyncHTTPTestCase):
     API_BASE = "/api/1.0"
 
     mock_runfolder_repo = MagicMock()
+    return_projects = True
 
     def get_app(self):
         self.mock_runfolder_repo.get_runfolders.return_value = FAKE_RUNFOLDERS
         self.mock_runfolder_repo.get_runfolder.return_value = FAKE_RUNFOLDERS[0]
+        def get_projects_from_runfolders():
+            if self.return_projects:
+                projs = []
+                for runfolder in FAKE_RUNFOLDERS:
+                    for project in runfolder.projects:
+                        projs.append(project)
+                return projs
+            else:
+                return []
+        self.mock_runfolder_repo.get_projects = get_projects_from_runfolders
 
         return Application(
             routes(
@@ -35,10 +46,11 @@ class TestProjectHandlers(AsyncHTTPTestCase):
                 expected_result.append(project.__dict__)
 
         self.assertEqual(response.code, 200)
-        self.assertDictEqual(json.loads(response.body), {"projects": expected_result})
+        result = json.loads(response.body)
+        self.assertDictEqual(result, {"projects": expected_result})
 
     def test_get_projects_empty(self):
-        self.mock_runfolder_repo.get_runfolders.return_value = []
+        self.return_projects = False
 
         response = self.fetch(self.API_BASE + "/projects")
